+++ src/optimizer/associative.ts
@@ -7,18 +7,17 @@ import {
     AstValue,
     isValue,
 } from "../grammar/ast";
-import { evalBinaryOp } from "../interpreter";
+import { evalBinaryOp } from "../interpreterSemantics/standardSemantics";
+import { abs, sign } from "../interpreterSemantics/util";
 import { Value } from "../types/types";
 import { ExpressionTransformer, Rule } from "./types";
 import {
-    abs,
     checkIsBinaryOpNode,
     checkIsBinaryOp_With_RightValue,
     checkIsBinaryOp_With_LeftValue,
     extractValue,
     makeBinaryExpression,
     makeValueExpression,
-    sign,
 } from "./util";
 
 type TransformData = {
